# swagger-ui/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Coffee Supply Chain",
        "description": "Coffee Supply Chain Application API",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "localhost:4000",
    "basePath": "/",
...
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/country/{id}": {
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "description": "ID of Country that we want to find",
                    "type": "integer"
                }
            ],
            "get": {
                "tags": [
                    "Countries"
                ],
                "summary": "Get Country from the system",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Countries_id"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Countries"
                ],
                "parameters": [
                    {
                        "name": "country",
                        "in": "body",
                        "description": "Country that we want to update",
                        "schema": {
                            "$ref": "#/definitions/Countries_id"
                        }
                    }
                ],
                "summary": "Get all Country from the system",
                "responses": {
                    "201": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Countries_id"
                        }
                    }
                }
            }
        },
        "/region": {
            "get": {
                "tags": [
                    "Regions"
                ],
                "summary": "Get all Regions from the system",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Regions"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Regions"
                ],
                "parameters": [
                    {
                        "name": "region",
                        "in": "body",
                        "description": "Region that we want to create",
                        "schema": {
                            "$ref": "#/definitions/Regions_id"
                        }
                    }
                ],
                "summary": "Get the Region created",
                "responses": {
                    "201": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Regions_id"
                        }
                    }
                }
            }
        },
        "/region/country/{country_id}": {
            "get": {
                "tags": [
                    "Regions"
                ],
                "parameters": [
                    {
                        "name": "country_id",
                        "in": "path",
                        "required": true,
                        "description": "country that we want to select",
                        "schema": {
                            "$ref": "#/definitions/Regions_country_id"
                        }
                    }
                ],
                "summary": "Get all Regions from the system using country id",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Regions"
                        }
                    }
                }
            }
        },
        "/region/{id}": {
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "description": "ID of Country that we want to find",
                    "type": "integer"
                }
            ],
            "get": {
                "tags": [
                    "Regions"
                ],
                "summary": "Get Region from the system",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Regions_id"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Regions"
                ],
                "parameters": [
                    {
                        "name": "region",
                        "in": "body",
                        "description": "Region that we want to update",
                        "schema": {
                            "$ref": "#/definitions/Regions_id"
                        }
                    }
                ],
                "summary": "Get all Region from the system",
                "responses": {
                    "201": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Regions_id"
                        }
                    }
                }
            }
        },
        "/producer": {
            "get": {
                "tags": [
                    "Producers"
                ],
                "summary": "Get all Producers from the system",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Producers"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Producers"
                ],
                "parameters": [
                    {
                        "name": "producer",
                        "in": "body",
                        "description": "Producer that we want to update",
                        "schema": {
                            "$ref": "#/definitions/Producers_id"
                        }
                    }
                ],
                "summary": "Get all Producers from the system",
                "responses": {
                    "201": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Producers_id"
                        }
                    }
                }
            }
        },
        "/producer/{id}": {
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "description": "ID of producer that we want to find",
                    "type": "integer"
                }
            ],
            "get": {
                "tags": [
                    "Producers"
                ],
                "summary": "Get all Producers from the system",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Producers_id"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Producers"
                ],
                "parameters": [
                    {
                        "name": "producer",
                        "in": "body",
                        "description": "Producer that we want to update",
                        "schema": {
                            "$ref": "#/definitions/Producers_id"
                        }
                    }
                ],
                "summary": "Get all Producers from the system",
                "responses": {
                    "201": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Producers_id"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Producers": {
            "properties": {
                "_id": {
                    "type": "integer",
                    "uniqueItems": true
                },
                "isPublic": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "books": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "name": {
                                "type": "string"
                            },
                            "amount": {
                                "type": "number"
                            }
                        }
                    }
                },
                "companies": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "Producers_id": {
            "properties": {
                "name": {
                    "type": "string"
                },
                "address": {
                    "type": "string"
                },
                "region_id": {
                    "type": "integer"
                }
            }
        },
...
        "Countries_id": {
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "Regions": {
            "properties": {
                "id": {
                    "type": "integer",
                    "uniqueItems": true
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "Regions_id": {
            "properties": {
                "name": {
                    "type": "string"
                },
                "country_id": {
                    "type": "integer"
                }
            }
        },
        "Regions_country_id": {
            "properties": {
                "country_id": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "bearerAuth": {
          "name": "Authorization",
          "in": "header",
          "type": "apiKey",
          "description": "JWT Authorization header"
        }
      },
      "security": [ { "bearerAuth": [] } ]
}